+++ src/utils/markdown/prettifyMarkdown.ts
@@ -7,6 +7,7 @@ import type { string_html } from '../../types/typeAliases';
  *
  * @param content raw html code
  * @returns formatted html code
+ * @private withing the package because of HUGE size of prettier dependency
  */
 export function prettifyMarkdown<TContent extends string_html>(content: TContent): TContent {
     try {